requirements.txt
Django==2.2
django-autoslug==1.9.4
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
Pillow==6.0.0
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
